pyproject.toml
[tool.poetry]
name = "django_project"
version = "0.1.0"
description = "Screening project"
authors = ["marts-dev <8363978+marts-dev@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = "^3.8"
django = "^3.1"
djangorestframework = "^3.11.1"
django-environ = "^0.4.5"
drf-yasg = "^1.17.1"

[tool.poetry.dev-dependencies]
black = "^19.10b0"
mypy = "0.770"
pytest-django = "^3.9.0"
django-stubs = "^1.5.0"
djangorestframework-stubs = "^1.2.0"
drf-yasg-stubs = "^0.1.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
